% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization.R
\name{bcClusters}
\alias{bcClusters}
\title{Plots a UMAP reduction coloured by metadata information}
\usage{
bcClusters(bc, idents, UMAP = "beyondcell", factor.col = TRUE, ...)
}
\arguments{
\item{bc}{\code{\link[beyondcell]{beyondcell}} object.}

\item{idents}{Name of the metadata column to colour by.}

\item{UMAP}{UMAP reduction to plot. Either \code{"beyondcell"}, computed
using \code{\link[beyondcell]{bcUMAP}}, or \code{"Seurat"}, obtained using
\code{Seurat}'s functions.}

\item{factor.col}{Logical indicating if \code{idents} column is a factor or
not. Set \code{factor.col = FALSE} if \code{idents} is a numeric column (such
as \code{percent.mt} or \code{nFeature_RNA}).}

\item{...}{Other arguments passed to \code{\link[Seurat]{DimPlot}}.}
}
\value{
A \code{ggplot2} object with the UMAP reduction coloured by
\code{idents}.
}
\description{
This function returns a {\link[ggplot2]{ggplot2}} object with
a UMAP reduction coloured by the specified metadata column.
}
