% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Manipulation.R
\name{bcSubset}
\alias{bcSubset}
\title{Subsets a beyondcell object}
\usage{
bcSubset(
  bc,
  signatures = NULL,
  bg.signatures = NULL,
  cells = NULL,
  nan.sigs = 1,
  nan.cells = 1
)
}
\arguments{
\item{bc}{\code{beyondcell} object.}

\item{signatures}{Vector with the names of the signatures to subset by. If
\code{signatures = NULL}, all signatures will be kept.}

\item{bg.signatures}{Vector with the names of the background signatures to
subset by. If \code{bg.signatures = NULL}, all background signatures will be
kept.}

\item{cells}{Vector with the names of the cells to subset by. If
\code{cells = NULL}, all cells will be kept.}

\item{nan.sigs}{Maximum proportion of \code{NaN} values per signature in the
output \code{beyondcell} object. All signatures with a proportion of
\code{NaN > nan.sig} will be removed.}

\item{nan.cells}{Maximum proportion of \code{NaN} values per cell in the
output \code{beyondcell} object. All cells with a proportion of
\code{NaN > nan.cells} will be removed.}
}
\value{
A subsetted \code{beyondcell} object.
}
\description{
This function subsets a \code{\link[beyondcell]{beyondcell}}
object based on signature names, cells and/or the maximum proportion of
\code{NaN} values desired in each signature and/or cell. See Details for more
information.
}
\details{
This function can subset a \code{beyondcell} object using its 5
parameters alone or in combination.

So, for example, if you specify \code{signatures} and \code{cells}, the
resulting \code{beyondcell} object (except the \code{@background} slot) will
be subsetted according to those vectors. The slot \code{@background} will be
only subsetted according to \code{cells}. If you want to subset it by
signatures as well, you must specify a value for \code{bg.signatures}.

On the other hand, if you specify \code{cells} and \code{nan.sigs}, the
output \code{beyondcell} object will keep the selected cells and those
signatures with a proportion of \code{NaN} values below or equal to
\code{nan.sigs}. Note that \code{nan.sigs} and \code{nan.cells} arguments
also subset the signatures and cells that meet the criteria in
\code{@background} slot.

Finally, if you specify all parameters, the result will keep those signatures
and cells of interest with a proportion of \code{NaN} below or equal to
\code{nan.sigs} and \code{nan.cells}, respectively.
}
