% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reductions.R
\name{bcUMAP}
\alias{bcUMAP}
\title{Computes a UMAP projection and therapeutic clusters using the BCS}
\usage{
bcUMAP(
  bc,
  pc = NULL,
  k.neighbors = 20,
  res = 0.2,
  add.DSS = FALSE,
  method = "uwot",
  return.model = FALSE,
  elbow.path = NULL
)
}
\arguments{
\item{bc}{\code{\link[beyondcell]{beyondcell}} object.}

\item{pc}{Number of principal components to use (which can be estimated from
an elbow plot). If \code{pc = NULL} (default), the function will stop prior
to compute the UMAP projection and the therapeutic clusters. See Details for
more information.}

\item{k.neighbors}{(\code{\link[Seurat]{FindNeighbors}}' \code{k.param})
Defines \emph{k} for the k-Nearest Neighbour algorithm.}

\item{res}{(\code{\link[Seurat]{FindClusters}}' \code{resolution}) Value of
the resolution parameter, use a value above/below 1.0 if you want to obtain
a larger/smaller number of communities. Can be a single number or a numeric
vector.}

\item{add.DSS}{Use background BCS computed with \code{DSS} signatures
(\code{add.DSS = TRUE}) or just use the signatures included in the \code{bc}
object (\code{add.DSS = FALSE}) to compute the UMAP projection and the
therapeutic clusters. If the number of drugs in \code{bc}
(excluding pathways) is <= 20, it is recomended to set \code{add.DSS = TRUE}.
Note that if \code{add.DSS = TRUE}, the regression and subset steps that have
been applied on \code{bc} will also be applied on the background BCS.}

\item{method}{(\code{\link[Seurat]{RunUMAP}}'s \code{umap.method}) UMAP
implementation to run. Can be:
\itemize{
\item{\code{uwot}}: Runs the UMAP via the \code{\link[uwot]{umap}} function
of the \code{uwot} package.
\item{\code{umap-learn}}: Runs the UMAP via the \code{Seurat} wrapper of the
python \code{umap-learn} package.
}}

\item{return.model}{(\code{\link[Seurat]{RunUMAP}}'s \code{return.model})
Whether \code{RunUMAP} will return the \code{uwot} model.}

\item{elbow.path}{Path to save the elbow plot. If \code{elbow.path = NULL}
(default), the plot will be printed.}
}
\value{
A \code{beyondcell} object with the UMAP reduction in
\code{@reductions} slot and the therapeutic clusters for each \code{res}
in \code{bc@meta.data}. Also, an elbow plot (\code{\link[ggplot2]{ggplot2}}
object) is printed (if \code{elbow.path = NULL}) or saved (if
\code{elbow.path != NULL}).
}
\description{
This function uses the beyondcell scores (BCS) to compute a
UMAP projection of the data and to cluster cells according to their
sensitivity to the tested drugs (therapeutic clusters).
}
\details{
This function performs all the steps required to obtain a UMAP
reduction of the data and cluster the cells according to the BCS.

You will normally require to run the function twice:
\enumerate{
\item Using \code{pc = NULL} to obtain the elbow plot.
\item Specifying a value for the \code{pc} parameter according to this plot.
This second time, the UMAP reduction and the therapeutic clusters will be
computed.}

Note that \code{add.DSS} must be the same in both runs, so the elbow plot
obtained in 1 is still valid in 2. If \code{add.DSS = TRUE}, the background
BCS will be stored in the \code{bc} object and the function will skip this
step the second time.
}
