% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\docType{class}
\name{beyondcell-class}
\alias{beyondcell-class}
\alias{beyondcell}
\title{Beyondcell class}
\description{
An object to represent the beyondcell scores (BCS) for each cell and
signature.
}
\section{Slots}{

\describe{
\item{\code{scaled}}{(Subsetted and/or regressed) scaled BCS.}

\item{\code{normalized}}{(Subsetted and/or regressed) normalized BCS.}

\item{\code{data}}{Original normalized BCS, without subsetting or regression.}

\item{\code{switch.point}}{(Subsetted and/or regressed) scaled BCS for which the
normalized score in \code{@data} is 0 (one switch point per signature).}

\item{\code{ranks}}{List of dataframes with the BCS' statistics and ranks returned
by \code{\link[beyondcell]{bcRanks}}.}

\item{\code{expr.matrix}}{Single-cell expression matrix used to compute the BCS.}

\item{\code{meta.data}}{Dataframe that contains information about each cell
(including the therapeutic clusters and \code{\link[Seurat]{Seurat}}'s
\code{@meta.data}).}

\item{\code{SeuratInfo}}{List with information about the input \code{Seurat} object,
including the \code{@reductions}.}

\item{\code{background}}{(Subsetted and/or regressed) normalized BCS obtained using
DSS signatures. Useful to compute \code{beyondcell}'s UMAP reduction and the
therapeutic clusters when the number of drug signatures is low.}

\item{\code{reductions}}{A list of dimensional reductions for this object.}

\item{\code{regression}}{A list with the order of subset and regression steps
performed on the \code{beyondcell} object and the variables used for
regression.}

\item{\code{n.genes}}{Argument passed to \code{\link[beyondcell]{GenerateGenesets}}.
Number of up and/or down-regulated genes per signature.}

\item{\code{mode}}{Argument passed to \code{GenerateGenesets}. Whether the
\code{geneset} contains up and/or down-regulated genes.}

\item{\code{thres}}{Argument \code{expr.thres} passed to
\code{\link[beyondcell]{bcScore}}. Minimum fraction of signature genes that
must be expressed in a cell to compute its BCS.}
}}

