% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Format.R
\name{BreakString}
\alias{BreakString}
\title{Breaks a string into several lines}
\usage{
BreakString(x, split = ", ", line.length = 50)
}
\arguments{
\item{x}{String to be broken, formed by elements separated by \code{split}.}

\item{split}{Character that separates the elements of \code{x}.}

\item{line.length}{Length of the lines into which \code{x} will be broken.}
}
\value{
A string with the same content as \code{x} broken in lines of
length \code{line.length}.
}
\description{
This function breaks a string \code{x} formed by elements
separated by \code{split} into lines of length \code{line.length}.
}
