% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Basics.R
\name{colMinus}
\alias{colMinus}
\title{Computes the column substraction}
\usage{
colMinus(x, na.rm = FALSE)
}
\arguments{
\item{x}{A matrix or a dataframe.}

\item{na.rm}{(From \code{base}) Logical. Should missing values (including
\code{NaN}) from rows \code{2:length(x)} be omitted from the calculations?}
}
\value{
A numeric rectangular object with the result of the substraction.
}
\description{
This function substracts to the first element of each column of
a rectangular object (\code{x[1, n]}) the rest of elements of the same column
(\code{x[2:length(x), n]}).
}
