% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Manipulation.R
\name{CreatebcObject}
\alias{CreatebcObject}
\title{Creates a new beyondcell object}
\usage{
CreatebcObject(bc)
}
\arguments{
\item{bc}{\code{beyondcell} object.}
}
\value{
A new \code{beyondcell} object.
}
\description{
This function creates a new \code{\link[beyondcell]{beyondcell}}
object.
}
\details{
This function creates a new \code{beyondcell} object by using the
normalized BCS as the original \code{@data}. Switch points are
recomputed and \code{@regression} is restarted. The \code{@expr.matrix} and
\code{@meta.data} slots are subsetted to keep only those cells present in
the new \code{@data} slot.
}
