% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Genesets.R
\name{GenerateGenesets}
\alias{GenerateGenesets}
\title{Creates a geneset object}
\usage{
GenerateGenesets(
  x,
  n.genes = 250,
  mode = c("up", "down"),
  filters = list(drugs = NULL, IDs = NULL, MoAs = NULL, targets = NULL, sources = NULL),
  comparison = NULL,
  include.pathways = TRUE
)
}
\arguments{
\item{x}{A pre-loaded matrix, a ranked matrix or a path to a GMT file with
custom gene sets. See Details for more information.}

\item{n.genes}{Number of up and/or down-regulated genes used to compute each
signature.}

\item{mode}{Whether the output \code{geneset} must contain up and/or
down-regulated genes. See Details for more information.}

\item{filters}{If \code{x} is a pre-loaded matrix, you can provide a list of
filters to subset it. You can specify which \code{drugs}, sig \code{IDs},
mechanisms of action (\code{MoAs}), \code{targets} and/or \code{sources} you
are interested in (cap insensitive). You can call
\code{\link[beyondcell]{ListFilters}} to check all the available values for
these filters. The signatures that pass \strong{ANY} of them are included in
the output.}

\item{comparison}{\code{"treated_vs_control"} or
\code{"sensitive_vs_resistant"}. See Details for more information.}

\item{include.pathways}{Logical. Return \code{beyoncell}'s pre-computed
signatures for functional pathways?}
}
\value{
A \code{geneset} object.
}
\description{
This function creates a \code{\link[beyondcell]{geneset}}
object.
}
\details{
\code{x} can be:
\itemize{
\item{A pre-loaded matrix:} {Either \code{PSc}, \code{SSc} or \code{DSS}.}
\item{A ranked matrix:} {A matrix with genes as rows and signatures as
columns that contains some type of numeric value such a t-stat or a LFC to
rank the genes accordingly.}
\item{A path to a GMT file:} {A file that contains custom gene sets. Each
gene set must have an "_UP" or "_DOWN" suffix.}
}
In addition, \code{mode} can be:
\itemize{
\item{\code{"up"}:} {To compute the signatures using only up-regulated
genes.}
\item{\code{"down"}:} {To compute the signatures using only down-regulated
genes.}
\item{\code{c("up", "down")} :} {To compute the signatures using both up and
down-regulated genes.}
}
If \code{x} is a path to a GMT file, \code{mode} is deprecated and the names
of all gene sets must end in "_UP" or "_DOWN" to indicate the \code{mode} of
each one.

Finally, \code{comparison} can be:
\itemize{
\item{\code{"treated_vs_control"}:} {(\code{PSc} and \code{DSS} like) When
the numeric values or the gene sets in the GMT file were obtained from a
comparison between drug treated and untreated cells.}
\item{\code{"sensitive_vs_resistant"}:} {(\code{SSc} like) When the numeric
values or the gene sets in the GMT file were obtained from a comparison
between drug sensitive and resistant cells.}
}
When \code{x} is a pre-loaded matrix, \code{comparison} is set automatically.
}
