% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\docType{class}
\name{geneset-class}
\alias{geneset-class}
\alias{geneset}
\title{Geneset class}
\description{
An object to represent signatures.
}
\section{Slots}{

\describe{
\item{\code{genelist}}{A list of drug signatures and functional \code{pathways} (if
specified) with up and/or down-regulated genes.}

\item{\code{n.genes}}{Argument passed to \code{\link[beyondcell]{GenerateGenesets}}.
Number of up and/or down-regulated genes per signature.}

\item{\code{mode}}{Argument passed to \code{GenerateGenesets}. Whether the
\code{geneset} contains up and/or down-regulated genes.}

\item{\code{info}}{Dataframe with drug signatures information, including sig IDs,
drug names, MoAs, target genes and data sources (LINCS, CTRP, GDSC or CCLE).
This slot is only filled if \code{GenerateGenesets}' input is a pre-loaded
matrix.}

\item{\code{comparison}}{Argument passed to \code{GenerateGenesets}. Either
\code{"treated_vs_control"} or \code{"control_vs_treated"}.}
}}

