% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Format.R
\name{get_colour_steps}
\alias{get_colour_steps}
\title{Returns a vector of 5 colours}
\usage{
get_colour_steps(colorscale = NULL)
}
\arguments{
\item{colorscale}{Either a \code{viridis}, \code{RColorBrewer} or a custom
palette of 3 colours (low, medium and high). If \code{colorscale = NULL}
(default), the function returns \code{beyondcell}'s own palette.}
}
\value{
A vector with 5 colour values.
}
\description{
This function gets a colour palette and returns a vector with 5
colour values to be used in
\code{\link[beyondcell]{center_scale_colour_stepsn}}.
}
