% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Basics.R
\name{GetIDs}
\alias{GetIDs}
\title{Returns the IDs that match the specified filter's values}
\usage{
GetIDs(values, filter, df = drugInfo)
}
\arguments{
\item{values}{User-supplied filtering vector.}

\item{filter}{Column name or number to subset by.}

\item{df}{\code{data.frame} with drug information. It must contain, at least,
two columns: \code{"IDs"} and \code{filter}.}
}
\value{
A vector with the IDs that match the \code{filter}'s values.
}
\description{
This function subsets \code{df} to select only the entries that
match the specified \code{filter}'s \code{values} and returns the
corresponding IDs.
}
