% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Genesets.R
\name{ListFilters}
\alias{ListFilters}
\title{Returns all the possible values for the specified filter}
\usage{
ListFilters(entry)
}
\arguments{
\item{entry}{Either \code{"drugs"}, \code{"IDs"}, \code{"MoAs"},
\code{"targets"} or \code{"sources"}.}
}
\value{
All the possible values for the specified \code{entry}.
}
\description{
This function returns all the available values for \code{drugs},
\code{IDs}, \code{MoAs}, \code{targets} or \code{sources} filters in
\code{\link[beyondcell]{GenerateGenesets}} function.
}
