% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Basics.R
\name{Mean.Med.SD}
\alias{Mean.Med.SD}
\title{Computes the mean, the median and the sd of a vector}
\usage{
Mean.Med.SD(x)
}
\arguments{
\item{x}{Numeric vector.}

\item{na.rm}{(From base) Logical. Should missing values (including NaN) be
removed?}
}
\value{
A named numeric vector with the mean, median and sd of \code{x}.
}
\description{
This function computes the mean, the median and the sd of a
vector.
}
