% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Basics.R
\name{minus}
\alias{minus}
\title{Substracts to the first element of a vector the rest of elements}
\usage{
minus(x, na.rm = FALSE)
}
\arguments{
\item{x}{Numeric vector.}

\item{na.rm}{(From \code{base}) Logical. Should missing values (including
\code{NaN}) be removed?}
}
\value{
The result of the substraction.
}
\description{
This function substracts to the first element of a numeric
vector (\code{x[1]}) the rest of elements of the same vector.
(\code{x[2:length(x)]}).
}
