% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Score.R
\name{SwitchPoint}
\alias{SwitchPoint}
\title{Computes the switch point}
\usage{
SwitchPoint(bc)
}
\arguments{
\item{bc}{\code{beyondcell} object.}
}
\value{
A named vector with the swich points of the signatures in \code{bc}.
}
\description{
This function computes the switch point of the signatures of a
given \code{\link[beyondcell]{beyondcell}} object. The switch point is the
(subsetted and/or regressed) scaled beyondcell score (BCS) that corresponds
to the point in which the normalized BCS in \code{beyondcell@data} switchs
from negative (insensitive) to positive (sensitive) values. The closer to 0,
the more sensitive are the cells to a given drug.
}
